--- LoneMiner/Assets/OptionsMenuInventory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b3cf9db6739cd4e48a7d760d316553ed
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
