--- LoneMiner/Assets/Scripts/DotE.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fdf49cb7d605ea14a93f672f884fcfa1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
