+++ LoneMiner/Assets/Scripts/DotE.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 31884a3060f26ec4f9575ce1f53b5cc9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
