+++ LoneMiner/Assets/Scripts/DotE/DotEAsteroids.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 97b3a39010ba622429dcdbd350d157e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
