+++ LoneMiner/Assets/Scripts/DotE/DotEMissionController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b39050178d149d6418f15c81658ac8a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
