+++ LoneMiner/Assets/Scripts/HUD.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a6431dc02c325444cbe0103db6cdf903
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
