+++ LoneMiner/Assets/Scripts/InventoryMenuManager.cs
@@ -1,0 +1,1 @@
+﻿/* This script controls the persentation of a menu on a GUI canvas  * Use this as a base for your menu.  * The hot key is ESC.  * Pete Fennell 10-1-2018  */  using System.Collections; using System.Collections.Generic; using UnityEngine; using UnityEngine.UI;  public class InventoryMenuManager : MonoBehaviour {     public OptionsMenuInventory optionsMenu;      public Image[] itemImages = new Image[numItemSlots];     public Item[] items = new Item[numItemSlots];     public const int numItemSlots = 4;      // Use this for initialization     void Start()     {          optionsMenu.gameObject.SetActive(!optionsMenu.gameObject.activeSelf);      }      // Update is called once per frame     void Update()     {          if (Input.GetKeyDown(KeyCode.I))         {             optionsMenu.gameObject.SetActive(!optionsMenu.gameObject.activeSelf);             if (Input.GetKeyDown(KeyCode.I))             {                  if (Time.timeScale == 1)                      Time.timeScale = 0;                 else                     Time.timeScale = 1;             }          }      }     public void AddItem(Item itemToAdd)     {         for (int i = 0; i < items.Length; i++)         {             if (items[i] == null)             {                 items[i] = itemToAdd;                // itemImages[i].sprite = itemToAdd.sprite;                // itemImages[i].enabled = true;                 return;             }         }     }      public void RemoveItem(Item itemToRemove)     {         for (int i = 0; i < items.Length; i++)         {             if (items[i] == itemToRemove)             {                 items[i] = null;                 //itemImages[i].sprite = null;                 //itemImages[i].enabled = false;                 return;             }         }     } }
