+++ LoneMiner/Assets/Scripts/InventoryMenuManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e7e1ac52298cd6949b31b5c864de98d6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
