+++ LoneMiner/Assets/Scripts/NewBehaviourScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b90cc45246519a43be20d60ac6e10a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
